Create Padding I
In one line, set the .navigation li elements to have 20 pixels of padding. Click Run and observe their change.

2.
Look at the red boxes at the bottom of the web page. Set the .share a elements to have 14 pixels of padding in style.css and run your code. Observe how the red boxes at the bottom of the page changed.

3.
Set the top and bottom padding of h2 elements to 20 pixels and set the left and right padding of h2 elements to 30 pixels.