Create Hex and RGB
1.
In the .green rule in style.css, change the background color to #9EB599.

Run the code. Can you tell the difference?

2.
In the .light rule in style.css, change the background color to #683C2C.

3.
In the .city rule in style.css, change the background color to #4C352D.

4.
In the .vienna rule in style.css, change the background color to #352926.

5.
In the .italian rule in style.css, change the background color to #141212.

Run the code one more time. These new colors are a lot closer to the real-life color of each type of coffee. How does the subtle difference feel?